added GridViewPager
