+++ wear/src/main/AndroidManifest.xml
@@ -19,6 +19,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".NewTimerActivity"/>
         <service android:name=".TimerNotificationService"/>
     </application>
