+++ wear/src/main/java/com/ni032mas/poti/SettingsWearableAdapter.java
@@ -1,0 +1,54 @@
+package com.ni032mas.poti;
+
+import android.content.Context;
+import android.support.wearable.view.WearableListView;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.TextView;
+
+import java.util.Arrays;
+import java.util.List;
+
+/**
+ * Created by ivansv on 08.11.2015.
+ */
+public class SettingsWearableAdapter extends WearableListView.Adapter {
+    private List<String> mItems;
+    private final LayoutInflater mInflater;
+
+    public SettingsWearableAdapter(Context context, String[] items) {
+        this(context, Arrays.asList(items));
+    }
+    public SettingsWearableAdapter(Context context, List<String> items) {
+        mInflater = LayoutInflater.from(context);
+        mItems = items;
+    }
+
+    private static class ItemViewHolder extends WearableListView.ViewHolder {
+        private TextView mItemTextView;
+
+        public ItemViewHolder(View itemView) {
+            super(itemView);
+            mItemTextView = (TextView) itemView.findViewById(R.id.settings_name);
+        }
+    }
+
+    @Override
+    public WearableListView.ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
+        return new ItemViewHolder(mInflater.inflate(R.layout.settings_wlistview_item, null));
+    }
+
+    @Override
+    public void onBindViewHolder(WearableListView.ViewHolder viewHolder, int position) {
+        ItemViewHolder itemViewHolder = (ItemViewHolder) viewHolder;
+        TextView textView = itemViewHolder.mItemTextView;
+        textView.setText(mItems.get(position));
+        ((ItemViewHolder) viewHolder).itemView.setTag(position);
+    }
+
+    @Override
+    public int getItemCount() {
+        return mItems.size();
+    }
+}
