+++ wear/src/main/java/com/ni032mas/poti/TimerNotificationService.java
@@ -61,7 +61,7 @@ public class TimerNotificationService extends IntentService {
     }
 
     private void restartAlarm() {
-        Intent dialogIntent = new Intent(this, MainActivity.class);
+        Intent dialogIntent = new Intent(this, NewTimerActivity.class);
         dialogIntent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
         startActivity(dialogIntent);
         if (Log.isLoggable(TAG, Log.DEBUG)) {
