+++ wear/src/main/res/layout/activity_new_timer.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<android.support.wearable.view.WatchViewStub
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools" android:id="@+id/new_watch_view_stub"
+    android:layout_width="match_parent" android:layout_height="match_parent"
+    app:rectLayout="@layout/rect_activity_new_timer" app:roundLayout="@layout/round_activity_new_timer"
+    tools:context=".NewTimerActivity"
+    tools:deviceIds="wear"></android.support.wearable.view.WatchViewStub>
