+++ wear/src/main/res/layout/round_activity_new_timer.xml
@@ -3,7 +3,5 @@
     xmlns:tools="http://schemas.android.com/tools" android:layout_width="match_parent"
     android:layout_height="match_parent" tools:context=".MainActivity" tools:deviceIds="wear_round">
 
-    <android.support.wearable.view.GridViewPager android:layout_height="match_parent"
-        android:layout_width="match_parent"
-        android:id="@+id/pager"/>
+    <include layout="@layout/timer_set_timer"/>
 </RelativeLayout>
