+++ wear/src/main/res/layout/settings_wlistview_item.xml
@@ -1,0 +1,8 @@
+<TextView
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:text="NAME"
+    android:id="@+id/settings_name"
+    android:layout_alignParentTop="true"
+    android:layout_alignParentStart="true"
+    xmlns:android="http://schemas.android.com/apk/res/android" />
